- Player names are now being exchanged on sync, but there is some   confusion with the indexing.  I can get the correct data on one   machine, but not another at this point.  Just some kind of off-by-one   issue I guess....

+++ src/burner/win32/burner_win32.h
@@ -92,9 +92,9 @@ struct DirectConnectionOptions {
   std::string romName;
   std::string localAddr = "";
   std::string remoteAddr = "";
-  size_t playerNumber = 0;
+  UINT16 playerNumber = 0;
   std::string playerName = "";
-  size_t frameDelay = 1;
+  UINT16 frameDelay = 1;
 };
